api routers added
Moved user api to a dedicate router\nPost user fixed
