+++ sitesManagementAPI/src/crud.py
@@ -17,7 +17,7 @@ def get_users(db: Session, skip: int = 0, limit: int = 100):
 
 
 def create_user(db: Session, user: schemas.BaseUser):
-    db_user = models.User(email=user.email, address=user.address)
+    db_user = models.User(**user.dict())
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
